extract query builder
