+++ Runtime/Api/Common/QueryBuilder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1329348e8031ff4a8db2e36e4da6b6d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
